Fundamental Functionality Recovery
1. Add Status item
2. Delete unused code
3. Change adjustable size of wallpaper cover in explorer context menu
4. Change horizontal layout constraint logic of wallpaper explorer
5. New PlayingInfo toolbar item on top center of wallpaper explorer window
6. Modify some convenience initializers of NSMenuItem

+++ Wallpaper Player/ContentView/Components/ContextMenus/ExplorerGlobalMenu.swift
@@ -27,9 +27,9 @@ struct ExplorerGlobalMenu: SubviewOfContentView {
             Menu("View") {
                 Section {
                     Picker("Icon Size", selection: $viewModel.explorerIconSize) {
-                        Text("Small Icons").tag(Double(100))
-                        Text("Medium Icons").tag(Double(125))
-                        Text("Large Icons").tag(Double(150))
+                        Text("Small Icons").tag(Double(150))
+                        Text("Medium Icons").tag(Double(200))
+                        Text("Large Icons").tag(Double(250))
                     }
                     .pickerStyle(.inline)
                 }
